Fixed some relevant warning on windows.
git-svn-id: svn://falconpl.org/falcon/core/trunk@1212 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/item_co.cpp
@@ -580,7 +580,7 @@ void co_string_div( const Item& first, const Item& second, Item& third )
             if ( chr >= -(int64) 0xFFFFFFFF && chr <= (int64) 0xFFFFFFFF && len > 0 )
             {
                CoreString *gcs = new CoreString( *str );
-               gcs->setCharAt( len-1, gcs->getCharAt(len-1) + chr );
+               gcs->setCharAt( len-1, gcs->getCharAt(len-1) + (uint32)chr );
                third = gcs;
                return;
             }
